requirements.txt
python-telegram-bot==21.6
openai>=1.75.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.5
python-docx>=1.1.2
